feat: auth api 추가 및 api 폴더 구조 재구성
