+++ src/api/apiClient.ts
@@ -1,0 +1,6 @@
+import axios from 'axios';
+
+export const apiClient = axios.create({
+  baseURL: 'https://linkbrary-api.vercel.app/13-6',
+  headers: { 'Content-Type': 'application/json' },
+});
